style: adds responsiveness to carousel slide 1 and better organizes hero carousel code into react components
