+++ src/views/Home.jsx
@@ -1,5 +1,5 @@
 // LIBRARY IMPORTS
-import React from 'react';
+import React, { useRef } from 'react';
 import { Link } from 'react-router-dom';
 
 // LOCAL IMPORTS
@@ -8,10 +8,12 @@ import StoryMission from '../components/StoryMission.jsx';
 import SingleCard from '../components/SingleCard.jsx';
 
 export function Home({ fullDataSet, changeVoicesColor,resetData}) {
+  // HANDLERS & CONSTANTS
   const limitData = []
   for(let i = 0; i<3; i++){
     limitData.push(fullDataSet[i])
   };
+
   const limitDataCard = limitData.map(data => {
     return (
       <SingleCard
@@ -23,14 +25,21 @@ export function Home({ fullDataSet, changeVoicesColor,resetData}) {
       />
     );
   });
+
   const handleVoiceIconClick = () => {
     resetData()
     changeVoicesColor();
   };
+
+  const firstItem = useRef(null);
+  const scroll = () =>{
+    firstItem.current.scrollIntoView();
+  }
+
   return (
     <div className="bg-base-100">
-   
-      <StoryMission />
+      <HeroCarousel scroll = {scroll} />
+      <StoryMission firstItem = {firstItem} />
       <div className="max-w-[1200px] xl:mx-auto md:mx-8 sm:mx-8">
         <div className="lg:text-3xl text-2xl font-bold text-center text-primary px-4">
             Explore Unmatched's Empowering Voices
@@ -40,7 +49,7 @@ export function Home({ fullDataSet, changeVoicesColor,resetData}) {
           <div className="grid lg:grid-cols-3 md:grid-cols-2 md:m-auto sm:grid-cols-1 gap-12 place-items-center m-auto">
         {limitDataCard}
         </div>
-        
+
         </div>
         <Link
               to="/professionals"
@@ -49,7 +58,7 @@ export function Home({ fullDataSet, changeVoicesColor,resetData}) {
              Unmatched Voices
             </Link>
         </div>
-       
+
       </div>
     </div>
   )
